+++ commhandler/src/sgr_commhandler/driver/modbus/modbus_client_async.py
@@ -30,9 +30,10 @@ class SGrModbusClient(ABC):
         self._lock = threading.Lock()
         self._client: ModbusBaseClient = client
         self._byte_order: Endian = (
-            Endian.BIG
-            if endianness is None or endianness == BitOrder.BIG_ENDIAN
-            else Endian.LITTLE
+            Endian.LITTLE
+            if endianness
+            in {BitOrder.CHANGE_BYTE_ORDER, BitOrder.CHANGE_BIT_ORDER}
+            else Endian.BIG
         )
         self._word_order: Endian = (
             Endian.LITTLE